tools/speakers/requirements.txt
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
cffi==1.13.2
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
entrypoints==0.3
flake8==3.7.9
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.2
mccabe==0.6.1
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.1.1
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
PyYAML==5.2
queuelib==1.5.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.13.0
Twisted==19.10.0
w3lib==1.21.0
zope.interface==4.7.1
